requirement.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
